requirements.txt
# --- GraphRAG core ---
graphrag>=0.3.6

# --- Azure OpenAI ---
openai>=1.30.0
tiktoken>=0.7.0
azure-identity>=1.15.0
azure-core>=1.30.0

# --- LangChain (for dossier chain integration) ---
langchain>=0.2.0
langchain-openai>=0.1.0

# --- Data handling ---
pandas>=2.0.0
numpy>=1.26.0
PyYAML>=6.0.0
python-dotenv>=1.0.0
pydantic>=2.5.0

# --- Async / HTTP ---
httpx>=0.27.0
aiofiles>=23.0.0
tenacity>=8.0.0

# --- Utilities ---
python-json-logger>=2.0.0
structlog>=24.0.0